+++ Tests/Reqnroll.SystemTests/Plugins/PluginsTestBase.cs
@@ -1,0 +1,5 @@
+﻿namespace Reqnroll.SystemTests.Plugins;
+
+public abstract class PluginsTestBase : SystemTestBase
+{
+}